+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added access control to tenant in myems-api, myems-admin and myems-web
+- added point real time report to myems-api
 - 
 ### Changed
 -